refactor: DTO 패키지를 service 하위에서 controller 하위로 이동
- 계층 의존 방향은 맞지 않지만, DTO의 실제 역할을 더 명확히 드러내는 방향의 피드백 반영
