+++ src/main/java/roomescape/reservation/controller/dto/request/ReservationCreateRequest.java
@@ -1,11 +1,12 @@
-package roomescape.wating.service.dto.request;
+package roomescape.reservation.controller.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotNull;
+
 import java.time.LocalDate;
 
-public record WaitingCreateRequest(
+public record ReservationCreateRequest(
         @NotBlank(message = "이름을 입력해야 합니다.")
         String name,
         @NotBlank(message = "이메일을 입력해야 합니다.")
