+++ src/main/java/roomescape/wating/controller/dto/request/WaitingCreateRequest.java
@@ -1,12 +1,11 @@
-package roomescape.reservation.service.dto.request;
+package roomescape.wating.controller.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotNull;
-
 import java.time.LocalDate;
 
-public record ReservationCreateRequest(
+public record WaitingCreateRequest(
         @NotBlank(message = "이름을 입력해야 합니다.")
         String name,
         @NotBlank(message = "이메일을 입력해야 합니다.")
