+++ src/main/java/roomescape/wating/controller/dto/response/WaitingCreateResponse.java
@@ -1,4 +1,4 @@
-package roomescape.wating.service.dto.response;
+package roomescape.wating.controller.dto.response;
 
 public record WaitingCreateResponse(
         Long id
